SvelteKit skeleton project
Having hammered out the basics of a newsletter archive it seemed best to keep it separate from the newsletter generator. This commit lays the foundations with a no-frills skeleton project from SvelteKit, which the first iteration of the archive will be built with. It has Netlify deployment set up too (hopefully) so the project can be live and shareable we work on it.